+++ app/charts/bar/chart-bar.tsx
@@ -134,12 +134,14 @@ const ChartBars = memo((props: ChartProps<BarConfig>) => {
           </ChartContainer>
           {fields.animation || limits.limits.length > 0 ? (
             <ChartControlsContainer>
-              <LegendColor
-                limits={limits.limits}
-                dimensionsById={dimensionsById}
-                chartConfig={chartConfig}
-                symbol="square"
-              />
+              {limits.limits.length > 0 && (
+                <LegendColor
+                  limits={limits.limits}
+                  dimensionsById={dimensionsById}
+                  chartConfig={chartConfig}
+                  symbol="square"
+                />
+              )}
               {fields.animation && (
                 <TimeSlider
                   filters={filters}