+++ app/charts/combo/combo-line-single-state.tsx
@@ -15,7 +15,9 @@ import {
   useYScales,
 } from "@/charts/combo/combo-state";
 import {
+  AxisLabelSizeVariables,
   getChartWidth,
+  useAxisLabelSizeVariables,
   useChartBounds,
   useChartPadding,
 } from "@/charts/shared/chart-dimensions";
@@ -50,6 +52,8 @@ export type ComboLineSingleState = CommonChartState &
     getColorLabel: (label: string) => string;
     chartWideData: ArrayLike<Observation>;
     getAnnotationInfo: (d: Observation) => TooltipInfo;
+    leftAxisLabelSize: AxisLabelSizeVariables;
+    bottomAxisLabelSize: AxisLabelSizeVariables;
   };
 
 const useComboLineSingleState = (
@@ -58,7 +62,7 @@ const useComboLineSingleState = (
   data: ChartStateData
 ): ComboLineSingleState => {
   const { chartConfig, measuresById } = chartProps;
-  const { xDimension, getX, getXAsString } = variables;
+  const { xDimension, getX, getXAsString, xAxisLabel } = variables;
   const { chartData, scalesData, timeRangeData, paddingData, allData } = data;
   const { fields, interactiveFiltersConfig } = chartConfig;
 
@@ -106,6 +110,7 @@ const useComboLineSingleState = (
 
   // Dimensions
   const { left, bottom } = useChartPadding({
+    xLabelPresent: !!xAxisLabel,
     yScale: paddingYScale,
     width,
     height,
@@ -113,6 +118,18 @@ const useComboLineSingleState = (
     formatNumber,
   });
   const margins = getMargins({ left, bottom });
+  const leftAxisLabelSize = useAxisLabelSizeVariables({
+    label: yAxisLabel,
+    width,
+    marginLeft: margins.left,
+    marginRight: margins.right,
+  });
+  const bottomAxisLabelSize = useAxisLabelSizeVariables({
+    label: xAxisLabel,
+    width,
+    marginLeft: margins.left,
+    marginRight: margins.right,
+  });
   const chartWidth = getChartWidth({
     width,
     left: margins.left,
@@ -180,6 +197,8 @@ const useComboLineSingleState = (
     getColorLabel: (label) => label,
     chartWideData,
     getAnnotationInfo,
+    leftAxisLabelSize,
+    bottomAxisLabelSize,
     ...variables,
   };
 };