+++ app/charts/shared/chart-state.ts
@@ -117,7 +117,10 @@ export const useChartState = () => {
 export type ChartWithInteractiveXTimeRangeState =
   | AreasState
   | ColumnsState
-  | LinesState;
+  | LinesState
+  | ComboLineSingleState
+  | ComboLineColumnState
+  | ComboLineDualState;
 
 export type ChartWithInteractiveYTimeRangeState = BarsState;
 
@@ -157,6 +160,7 @@ export const useBaseVariables = (chartConfig: ChartConfig): BaseVariables => {
 };
 
 export type BandYVariables = {
+  yAxisLabel: string;
   yDimension: Dimension;
   getY: StringValueGetter;
   getYLabel: (d: string) => string;
@@ -166,6 +170,7 @@ export type BandYVariables = {
 };
 
 export type BandXVariables = {
+  xAxisLabel: string;
   xDimension: Dimension;
   getX: StringValueGetter;
   getXLabel: (d: string) => string;
@@ -207,7 +212,10 @@ export const useBandYVariables = (
     dimensionsById[y.componentId].values
   )(y.componentId);
 
+  const yAxisLabel = getLabelWithUnit(yDimension);
+
   return {
+    yAxisLabel,
     yDimension,
     getY,
     getYLabel,
@@ -252,7 +260,10 @@ export const useBandXVariables = (
     dimensionsById[x.componentId].values
   )(x.componentId);
 
+  const xAxisLabel = getLabelWithUnit(xDimension);
+
   return {
+    xAxisLabel,
     xDimension,
     getX,
     getXLabel,
@@ -268,6 +279,7 @@ export type TemporalXVariables = {
   xDimension: TemporalDimension | TemporalEntityDimension;
   getX: TemporalValueGetter;
   getXAsString: StringValueGetter;
+  xAxisLabel: string;
 };
 
 export const useTemporalXVariables = (
@@ -294,7 +306,10 @@ export const useTemporalXVariables = (
   )(x.componentId);
   const getXAsString = useStringVariable(x.componentId);
 
+  const xAxisLabel = getLabelWithUnit(xDimension);
+
   return {
+    xAxisLabel,
     xDimension,
     getX: isTemporalDimension(xDimension) ? getXTemporal : getXTemporalEntity,
     getXAsString,